fix: long type for foreign key

+++ src/main/java/it/gov/pagopa/gpd/rtp/entity/PaymentOption.java
@@ -18,7 +18,7 @@ public class PaymentOption {
 
   @JsonProperty("payment_position_id")
   @Column(name = "payment_position_id")
-  private int paymentPositionId;
+  private Long paymentPositionId;
 
   private long amount;
   private String description;
